PID on floats instead of int32_t

+++ FlightController_v1_0/PID.cpp
@@ -9,7 +9,7 @@
 #include "PID.h"
 
 int32_t
-PID::get_pid(int32_t error)
+PID::get_pid(float error)
 {
     uint32_t tnow = millis();
     uint32_t dt = tnow - _last_t;
